rm unnecessary casts from GlobalSignal

+++ dioxus/src/components/global_signal/mod.rs
@@ -15,9 +15,9 @@ pub fn FnGlobalSignal() -> Element {
 
         Nullifier {}
 
-        Incrementer { increment_by: -1 as i8 }
-        Incrementer { increment_by: -5 as i8 }
-        Incrementer { increment_by: -10 as i8 }
-        Incrementer { increment_by: -20 as i8 }
+        Incrementer { increment_by: -1 }
+        Incrementer { increment_by: -5 }
+        Incrementer { increment_by: -10 }
+        Incrementer { increment_by: -20 }
     }
 }
